app/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.6
django-braces==1.15.0
django-cache-url==3.4.5
django-debug-toolbar==4.4.6
django-embed-video==1.4.9
django-redisboard==8.4.0
djangorestframework==3.15.2
environs==11.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
marshmallow==3.21.3
msgpack==1.0.8
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pymemcache==4.0.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==6.4.post2
